Added endpoints for getting and updating a todo Removed the todoAdd route Added a stacked button Updated the todo form to be edit only and to fetch the todo. Added ability to add todo from the todo list Added the click ouside directive Updated RepeatType to use shorts Updated TodoFormDto to return this from set methids In the TodoService, update does more of the heavy lifting Updated how jooq generates the field for bit(7)

+++ src/engine/src/main/java/space_habit_frontier/engine/constants/RepeatType.java
@@ -6,19 +6,19 @@ import space_habit_frontier.engine.interfaces.FriendlyNameable;
 import org.apache.commons.text.WordUtils;
 
 public enum RepeatType implements FriendlyNameable {
-	DATE(0),
-	DAILY(1),
-	WEEKLY(7),
-	MONTHLY(31),
-	YEARLY(366);
+	DATE((short)0),
+	DAILY((short)1),
+	WEEKLY((short)7),
+	MONTHLY((short)31),
+	YEARLY((short)366);
 	
-	private int __repeattype;
+	private short __repeattype;
 	
-	RepeatType(int repeattype) {
+	RepeatType(short repeattype) {
 		this.__repeattype = repeattype;
 	}
 	
-	public int getValue() {
+	public short getValue() {
 		return __repeattype;
 	}
 
@@ -26,7 +26,7 @@ public enum RepeatType implements FriendlyNameable {
 		return WordUtils.capitalizeFully(this.name());
 	}
 
-	public static RepeatType valueOf(int value) {
+	public static RepeatType valueOf(short value) {
 		return RepeatType.values()[value];
 	}
 }
